Change the product code to be a uint8_t consistently everwhere.
tic_settings_get_product was in a tagged release and it uses uint8_t, so we
should not change its signature and it's better to make everything else be
consistent with it.

tic_device_get_product has not been used in a tagged release yet, so it is OK to
change its signature.

+++ include/tic.h
@@ -1095,7 +1095,7 @@ void tic_device_free(tic_device *);
 /// Gets the product code of the device, which will be one of the TIC_PRODUCT_*
 /// macros.  If the device is NULL, returns 0 instead.
 TIC_API TIC_WARN_UNUSED
-uint32_t tic_device_get_product(const tic_device *);
+uint8_t tic_device_get_product(const tic_device *);
 
 /// Gets the full name of the device (e.g. "Tic USB Stepper Motor Controller
 /// T825") as an ASCII-encoded string.  The string will be valid at least as
